extracting contract
