+++ app/src/androidTest/java/br/com/rbrthmn/ui/financialcompanion/viewmodels/BalanceCardViewModelTest.kt
@@ -18,16 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.data.di
+package br.com.rbrthmn.ui.financialcompanion.viewmodels
 
-import br.com.rbrthmn.contracts.BalanceContract
-import br.com.rbrthmn.ui.financialcompanion.utils.DecimalFormatter
-import br.com.rbrthmn.ui.financialcompanion.viewmodels.BalanceCardViewModel
-import org.koin.androidx.viewmodel.dsl.viewModelOf
-import org.koin.dsl.module
-
-val appModule = module {
-    viewModelOf<BalanceContract.BalanceCardViewModel>(::BalanceCardViewModel)
-
-    factory { DecimalFormatter() }
+class BalanceCardViewModelTest {
 }
